Racist/backend/requirements.txt
flask>=2.2,<3.0
face-recognition
pillow
numpy
scikit-learn
